requirements.txt
Pillow>=12.0.0
pydantic
PyYAML>=6.0
Flask>=2.0.0
flask-cors>=4.0.0

# 生产环境可选依赖
gunicorn>=20.0.0